refactor(voting): group question validation and result into libs
