+++ packages/voting/graphql/resolvers/QuestionTypeRangeResult.js
@@ -1,7 +1,7 @@
-const { rangeResultHistogram } = require('../../lib/Question')
+const { resultRangeHistogram } = require('../../lib/Question')
 
 module.exports = {
   histogram (result, args, context) {
-    return rangeResultHistogram(result, args, context)
+    return resultRangeHistogram(result, args, context)
   }
 }
