+++ packages/voting/graphql/resolvers/_mutations/submitAnswer.js
@@ -4,7 +4,9 @@ const {
   ensureReadyToSubmit,
   transformQuestion
 } = require('../../../lib/Questionnaire')
-const { graphql: { resolvers: { queries: { document: getDocument } } } } = require('@orbiting/backend-modules-documents')
+const {
+  validateAnswer
+} = require('../../../lib/Question')
 
 module.exports = async (_, { answer: { questionId, payload } }, context) => {
   const { pgdb, user: me, t, req } = context
@@ -30,70 +32,15 @@ module.exports = async (_, { answer: { questionId, payload } }, context) => {
         throw new Error(t('api/questionnaire/answer/empty'))
       }
 
-      const { value } = payload
-      if (question.type === 'Text') {
-        if (typeof value !== 'string') {
-          throw new Error(t('api/questionnaire/answer/wrongType'))
-        }
-        if (value.length === 0) {
-          emptyAnswer = true
-        }
-      } else if (question.type === 'Range') {
-        if (typeof value !== 'number') {
-          throw new Error(t('api/questionnaire/answer/wrongType'))
-        }
-        const values = question.typePayload.ticks.map(t => t.value)
-        if (question.typePayload.kind === 'continous') {
-          const min = Math.min(...values)
-          const max = Math.max(...values)
-          if (value < min || value > max) {
-            throw new Error(t('api/questionnaire/answer/outOfRange'))
-          }
-        } else { // discrete
-          if (values.find(v => v === value) === undefined) {
-            throw new Error(t('api/questionnaire/answer/outOfRange'))
-          }
-        }
-      } else if (question.type === 'Choice') {
-        if (!Array.isArray(value)) {
-          throw new Error(t('api/questionnaire/answer/wrongType'))
-        }
-        if (value.length === 0) {
-          emptyAnswer = true
-        } else {
-          if (question.typePayload.cardinality > 0 && value.length > question.typePayload.cardinality) {
-            throw new Error(t('api/questionnaire/answer/Choice/tooMany', { max: question.typePayload.cardinality }))
-          }
-          for (let v of value) {
-            if (!question.typePayload.options.find(ov => ov.value === v)) {
-              throw new Error(t('api/questionnaire/answer/Choice/value/404', { value: v }))
-            }
-          }
-        }
-      } else if (question.type === 'Document') {
-        if (typeof value !== 'string') {
-          throw new Error(t('api/questionnaire/answer/wrongType'))
-        }
-        if (value.length === 0) {
-          emptyAnswer = true
-        } else {
-          const doc = await getDocument(null, { path: value }, context)
-          if (!doc) {
-            throw new Error(t('api/questionnaire/answer/Document/404', { path: value }))
-          }
-          const requestedTemplate = question.typePayload.template
-          if (requestedTemplate && requestedTemplate !== doc.meta.template) {
-            throw new Error(t('api/questionnaire/answer/Document/wrongTemplate', { template: requestedTemplate }))
-          }
-          // save doc to payload
-          payload.document = {
-            title: doc.meta.title,
-            credits: doc.meta.credits
-          }
-        }
-      } else {
-        throw new Error(t('api/questionnaire/question/type/404', { type: question.type }))
-      }
+      emptyAnswer = await validateAnswer(
+        payload.value,
+        question,
+        {
+          ...context,
+          pgdb: transaction
+        },
+        payload
+      )
     }
 
     const findQuery = {
