+++ packages/voting/lib/Question/Text.js
@@ -1,0 +1,13 @@
+const validate = (value, question, { t }) => {
+  if (typeof value !== 'string') {
+    throw new Error(t('api/questionnaire/answer/wrongType'))
+  }
+  if (value.length === 0) {
+    return true
+  }
+  return false
+}
+
+module.exports = {
+  validate
+}
